LayoutTests/http/tests/security/strip-referrer-to-origin-for-third-party-requests-in-private-mode-expected.txt

Tests that only the origin is sent as referrer for third-party resources in private browsing mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "http://127.0.0.1:8000/security/strip-referrer-to-origin-for-third-party-requests-in-private-mode.html"
PASS referrer is "http://127.0.0.1:8000/"
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
http://127.0.0.1:8000/
